+++ client/src/components/Auth/auth-form/components/index.tsx
@@ -1,0 +1,2 @@
+export {default as SignUp} from './sign-up.auth';
+export {default as SignIn} from './sign-in.auth';
