--- client/src/components/Auth/components/index.tsx
@@ -1,2 +1,0 @@
-export {default as signUpProps} from './sign-up.auth';
-export {default as signInProps} from './sign-in.auth';
