--- client/src/components/Auth/components/types.ts
@@ -1,14 +1,0 @@
-interface IInputField {
-  placeholder: string,
-  autoComplete?: string,  
-  label: string,
-  type?: string,
-  error?: string | undefined, 
-}
-
-export interface IInputFields {
-  username?: IInputField,
-  email: IInputField,
-  password: IInputField,
-  confirmPassword?: IInputField,
-}
