+++ server/src/apollo/resolvers/mutations/auth.mutations.js
@@ -1,0 +1,46 @@
+const {auth} = require('../../firebase-config');
+
+const setError = (code, message) => {
+  return ({
+    code,
+    message
+  })
+}
+
+exports.createUserWithEmailAndPassword = async (_, {email, password, username}) => {
+  const authenticatedUser = {
+    email,
+    password,
+    displayName: username,
+  }
+
+  try {
+    await auth().createUserWithEmailAndPassword(email, password)
+    const user = await auth().currentUser;
+
+    if (user) {
+      await user.updateProfile({
+        displayName: username,
+      }).catch(({code, message}) => {
+        authenticatedUser.error = setError(code, message)
+      });
+    } else {
+      authenticatedUser.error = setError('auth/no-current-user', 'user created with email but unable to update with username');
+    }
+
+  } catch({code, message}) {
+    authenticatedUser.error = setError(code, message)
+  }
+
+  return authenticatedUser;
+  // .then((userCredential) => {
+  //   // Signed in 
+  //   var user = userCredential.user;
+  //   // ...
+  // })
+  // .catch((error) => {
+  //   var errorCode = error.code;
+  //   var errorMessage = error.message;
+  //   // ..
+  // });
+}
