+++ server/src/apollo/resolvers/queries/index.js
@@ -1,0 +1,3 @@
+const trackQueries = require('./tracks.queries');
+
+module.exports = trackQueries;
